Convert more resources to custom resources and cleanup region handling
Convert everything that didn't do a load current resource to a custom resource
Move the region logic to a more predictable setup. Use resource's value first and set the default via the helper which pulls from Ohai or node attribute (when not actually in AWS)
