+++ .kitchen.yml
@@ -23,6 +23,20 @@ platforms:
     username: ubuntu
 
 suites:
+  - name: cfn_stack
+    run_list:
+    - recipe[aws_test::cfn_stack]
+    attributes:
+      aws_test:
+        key_id: <%= ENV['AWS_ACCESS_KEY_ID'] %>
+        access_key: <%= ENV['AWS_SECRET_ACCESS_KEY'] %>
+  - name: dynamo_table
+    run_list:
+    - recipe[aws_test::dynamodb_table]
+    attributes:
+      aws_test:
+        key_id: <%= ENV['AWS_ACCESS_KEY_ID'] %>
+        access_key: <%= ENV['AWS_SECRET_ACCESS_KEY'] %>
 
   - name: ebs_volume
     run_list:
