+++ resources/dynamodb_table.rb
@@ -8,7 +8,7 @@ attribute :local_secondary_indexes, kind_of: Array
 attribute :global_secondary_indexes, kind_of: Array
 attribute :provisioned_throughput, kind_of: Hash, required: true
 attribute :stream_specification, kind_of: Hash
-attribute :region, kind_of: String
+attribute :region, kind_of: String, default: lazy { aws_region }
 
 # aws credential attributes
 attribute :aws_access_key, kind_of: String
@@ -16,6 +16,8 @@ attribute :aws_secret_access_key, kind_of: String
 attribute :aws_session_token, kind_of: String
 attribute :aws_assume_role_arn, kind_of: String
 attribute :aws_role_session_name, kind_of: String
+
+include AwsCookbook::Ec2 # needed for aws_region helper
 
 private
