+++ resources/elastic_lb.rb
@@ -3,7 +3,7 @@ property :aws_secret_access_key, String
 property :aws_session_token,     String
 property :aws_assume_role_arn,   String
 property :aws_role_session_name, String
-property :region,                String
+property :region,                String, default: lazy { aws_region }
 property :name,                  String
 property :listeners,             Array
 property :security_groups,       Array
@@ -11,6 +11,8 @@ property :subnets,               Array # for VPC networking
 property :availability_zones,    Array # for classic networking
 property :tags,                  Array
 property :scheme,                Array
+
+include AwsCookbook::Ec2 # needed for aws_region helper
 
 action :register do
   target_lb = find_elb
@@ -62,7 +64,7 @@ action_class do
   def elb
     require 'aws-sdk'
     Chef::Log.debug('Initializing the ElasticLoadBalancing Client')
-    @elb ||= create_aws_interface(::Aws::ElasticLoadBalancing::Client)
+    @elb ||= create_aws_interface(::Aws::ElasticLoadBalancing::Client, new_resource.region)
   end
 
   def find_elb
