+++ resources/kinesis_stream.rb
@@ -1,13 +1,54 @@
-default_action :create
-actions :create, :delete
-
-attribute :stream_name, kind_of: String, name_attribute: true
-attribute :starting_shard_count, kind_of: Integer, required: true
-attribute :region, kind_of: String
+property :stream_name, String, name_property: true
+property :starting_shard_count, Integer, required: true
+property :region, String, default: lazy { aws_region }
 
 # aws credential attributes
-attribute :aws_access_key, kind_of: String
-attribute :aws_secret_access_key, kind_of: String
-attribute :aws_session_token, kind_of: String
-attribute :aws_assume_role_arn, kind_of: String
-attribute :aws_role_session_name, kind_of: String
+property :aws_access_key, String
+property :aws_secret_access_key, String
+property :aws_session_token, String
+property :aws_assume_role_arn, String
+property :aws_role_session_name, String
+
+include AwsCookbook::Ec2 # needed for aws_region helper
+
+action :create do
+  unless stream_exists?
+    converge_by("create Kinesis stream #{new_resource.stream_name}") do
+      kinesis.create_stream(
+        stream_name: new_resource.stream_name,
+        shard_count: new_resource.starting_shard_count
+      )
+    end
+  end
+end
+
+action :delete do
+  if stream_exists?
+    converge_by("delete Kinesis stream #{new_resource.stream_name}") do
+      kinesis.delete_stream(stream_name: new_resource.stream_name)
+    end
+  end
+end
+
+action_class do
+  include AwsCookbook::Ec2
+
+  def kinesis
+    require 'aws-sdk'
+
+    Chef::Log.debug('Initializing the Kinesis Client')
+    @kinesis ||= create_aws_interface(::Aws::Kinesis::Client, new_resource.region)
+  end
+
+  # does_stream_exist - logic for checking if the stream exists
+  def stream_exists?
+    resp = kinesis.describe_stream(stream_name: new_resource.stream_name)
+    if !resp.empty?
+      true
+    else
+      false
+    end
+  rescue ::Aws::Kinesis::Errors::ResourceNotFoundException
+    false
+  end
+end
