+++ test/fixtures/cookbooks/aws_test/recipes/dynamodb_table.rb
@@ -58,6 +58,8 @@ aws_dynamodb_table 'test-dynamodb-table-create' do
 end
 
 aws_dynamodb_table 'test-dynamodb-table-update' do
+  aws_access_key node['aws_test']['key_id']
+  aws_secret_access_key node['aws_test']['access_key']
   table_name 'test-dynamodb-table'
   retries 6
   retry_delay 10
