+++ test/fixtures/cookbooks/aws_test/recipes/resource_tag.rb
@@ -2,6 +2,6 @@ aws_resource_tag node['ec2']['instance_id'] do
   aws_access_key node['aws_test']['key_id']
   aws_secret_access_key node['aws_test']['access_key']
   tags('Name' => 'AWS Cookbook Test Node',
-       'Environment' => node.chef_environment)
+       'Environment' => 'test_kitchen')
   action :update
 end
