+++ test/integration/dynamodb_table/rspec/Gemfile
