+++ test/integration/dynamodb_table/rspec/dynamodb_table_update_spec.rb
