--- test/integration/dynamodb_table_create/rspec/Gemfile
@@ -1,4 +1,0 @@
-source 'https://rubygems.org'
-
-gem 'aws-sdk'
-gem 'rspec'
